+++ sipn_reanalysis_ingest/luigitasks/main.py
@@ -1,16 +1,42 @@
 import datetime as dt
 import shutil
+from typing import Union
 
 import luigi
 
+from sipn_reanalysis_ingest.constants.cfsr import CFSR_DAILY_TAR_ON_OR_AFTER
 from sipn_reanalysis_ingest.constants.date import (
     DEFAULT_PROCESSING_DAY,
     DEFAULT_PROCESSING_MONTH,
 )
 from sipn_reanalysis_ingest.constants.paths import DATA_UNTAR_DIR
-from sipn_reanalysis_ingest.luigitasks.convert import Grib2ToNcDaily, Grib2ToNcMonthly
+from sipn_reanalysis_ingest.luigitasks.convert import (
+    Grib2In5DailyTarToDailyNc,
+    Grib2InDailyAnd5DailyTarsToDailyNc,
+    Grib2InDailyTarToDailyNc,
+    Grib2ToMonthlyNc,
+)
 from sipn_reanalysis_ingest.util.date import date_range, month_range
 from sipn_reanalysis_ingest.util.log import logger
+
+
+def make_daily_processing_requirement(
+    *,
+    date: dt.date,
+) -> Union[
+    Grib2InDailyAnd5DailyTarsToDailyNc,
+    Grib2In5DailyTarToDailyNc,
+    Grib2InDailyTarToDailyNc,
+]:
+    if date == CFSR_DAILY_TAR_ON_OR_AFTER:
+        return Grib2InDailyAnd5DailyTarsToDailyNc(date=date)
+    elif date < CFSR_DAILY_TAR_ON_OR_AFTER:
+        return Grib2In5DailyTarToDailyNc(date=date)
+    elif date > CFSR_DAILY_TAR_ON_OR_AFTER:
+        return Grib2InDailyTarToDailyNc(date=date)
+
+    # TODO: How to convince Mypy that the code will never reach this line?
+    raise RuntimeError('This should not be reachable')
 
 
 # TODO: Move data from wip dir to final location. Change from WrapperTask to regular
@@ -23,7 +49,7 @@ class ProcessDateRange(luigi.WrapperTask):
 
     def requires(self):
         return [
-            Grib2ToNcDaily(date=date)
+            make_daily_processing_requirement(date=date)
             for date in date_range(self.start_date, self.end_date)
         ]
 
@@ -40,7 +66,7 @@ class ProcessMonthRange(luigi.WrapperTask):
 
     def requires(self):
         return [
-            Grib2ToNcMonthly(
+            Grib2ToMonthlyNc(
                 # TODO: Use `luigi.date_interval.Month`? Generate type stubs??
                 month=dt.date(year_month.year, year_month.month, 1),
             )