+++ sipn_reanalysis_ingest/test/test_cfsr.py
@@ -3,7 +3,7 @@ import datetime as dt
 import pytest
 
 from sipn_reanalysis_ingest.util.cfsr import (
-    _expected_6hourly_forecast_suffixes_for_date,
+    _expected_5daily_6hourly_forecast_suffixes_for_date,
 )
 
 
@@ -11,25 +11,25 @@ from sipn_reanalysis_ingest.util.cfsr import (
     'date,expected',
     [
         pytest.param(
-            dt.date(2020, 1, 1),
+            dt.date(2000, 1, 1),
             [
-                '2019123118.grb2',
-                '2020010100.grb2',
-                '2020010106.grb2',
-                '2020010112.grb2',
+                '1999123118.grb2',
+                '2000010100.grb2',
+                '2000010106.grb2',
+                '2000010112.grb2',
             ],
         ),
         pytest.param(
-            dt.date(2020, 12, 31),
+            dt.date(2000, 12, 31),
             [
-                '2020123018.grb2',
-                '2020123100.grb2',
-                '2020123106.grb2',
-                '2020123112.grb2',
+                '2000123018.grb2',
+                '2000123100.grb2',
+                '2000123106.grb2',
+                '2000123112.grb2',
             ],
         ),
     ],
 )
-def test__expected_6hourly_forecast_suffixes_for_date(date, expected):
-    actual = _expected_6hourly_forecast_suffixes_for_date(date)
+def test__expected_5daily_6hourly_forecast_suffixes_for_date(date, expected):
+    actual = _expected_5daily_6hourly_forecast_suffixes_for_date(date)
     assert actual == expected